requirements.txt
aiohttp
attrs
cached_property
cerberus
graphviz
html2text
huepy
jinja2
packaging
pip
python-dateutil
pyyaml
requests
tomlkit
tqdm
